+++ lib/veto.dart
@@ -1,4 +1,5 @@
 library veto;
 
 export 'data/models/base_view_model.dart';
+export 'data/models/base_view_model.dart';
 export 'data/enums/view_model_state.dart';