+++ lib/widgets/view_model_builder.dart
@@ -7,9 +7,11 @@ class ViewModelBuilder<T extends BaseViewModel> extends StatefulWidget {
     required T Function() viewModelBuilder,
     Object? Function()? argumentBuilder,
     Key? key,
+    final bool keepAlive = true,
   })  : _builder = builder,
         _viewModelBuilder = viewModelBuilder,
         _argumentBuilder = argumentBuilder,
+        _keepAlive = keepAlive,
         super(key: key);
 
   /// Builder method that builds the widget tree.
@@ -21,12 +23,14 @@ class ViewModelBuilder<T extends BaseViewModel> extends StatefulWidget {
   /// Builder method that provides the [BaseViewModel.initialise] with arguments.
   final dynamic Function()? _argumentBuilder;
 
+  final bool _keepAlive;
+
   @override
   ViewModelBuilderState<T> createState() => ViewModelBuilderState<T>();
 }
 
 class ViewModelBuilderState<T extends BaseViewModel>
-    extends State<ViewModelBuilder<T>> {
+    extends State<ViewModelBuilder<T>> with AutomaticKeepAliveClientMixin {
   /// The current [BaseViewModel].
   late final T _viewModel;
 
@@ -50,5 +54,11 @@ class ViewModelBuilderState<T extends BaseViewModel>
   }
 
   @override
-  Widget build(BuildContext context) => widget._builder(context, _viewModel);
+  Widget build(BuildContext context) {
+    super.build(context);
+    return widget._builder(context, _viewModel);
+  }
+
+  @override
+  bool get wantKeepAlive => widget._keepAlive;
 }